requirements.txt
accelerate==1.0.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
augly==1.0.0
blessed==1.20.0
blobfile==3.0.0
cachetools==5.5.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.6
comm==0.2.2
compressed-tensors==0.8.0
conllu==6.0.0
contourpy==1.3.1
cycler==0.12.1
datasets==2.20.0
debugpy==1.8.11
decorator==5.1.1
diffusers==0.31.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dlib==19.24.6
dnspython==2.6.1
einops==0.8.0
email_validator==2.2.0
emoji==2.14.0
exceptiongroup==1.2.1
executing==2.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.5.0
func-timeout==4.3.5
gguf==0.10.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.0
googleapis-common-protos==1.65.0
gpustat==1.1.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.27.0
idna==3.7
importlib_metadata==8.5.0
iniconfig==2.0.0
interegular==0.3.3
iopath==0.1.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lark==1.1.9
llvmlite==0.43.0
lm-format-enforcer==0.10.9
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.1
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.57.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.5
outlines==0.0.46
packaging==24.1
pandas==2.2.2
parso==0.8.4
partial-json-parser==0.2.1.1.post4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.27.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycryptodomex==3.20.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ray==2.30.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
stanza==1.9.2
starlette==0.37.2
statsmodels==0.14.3
sympy==1.13.1
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
tomli==2.0.1
torch==2.5.1
torchao==0.5.0
torchtune==0.3.0
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
transformers==4.47.0
triton==3.1.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
vllm==0.6.4.post1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
word2number==1.1
xformers==0.0.28.post3
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.2
